+++ qbraid/transpiler/conversions/__init__.py
@@ -14,21 +14,12 @@ quantum software program types.
 
 .. currentmodule:: qbraid.transpiler.conversions
 
-Functions
-----------
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   update_registered_conversions
-
-
 Submodules
 -----------
 
 .. autosummary::
    :toctree: ../stubs/
-   
+
    braket_ahs
    braket
    cirq
@@ -39,6 +30,7 @@ Submodules
    pyquil
    pytket
    qiskit
+   cudaq
 
 """
 import importlib
@@ -56,7 +48,7 @@ valid_combinations_cache = set()
 conversion_functions = []
 
 
-def update_registered_conversions() -> None:
+def _update_registered_conversions() -> None:
     """
     Dynamically update the list of conversion functions based on current
     NATIVE_REGISTRY and QPROGRAM_REGISTRY. Adds valid conversion functions
@@ -107,4 +99,4 @@ def update_registered_conversions() -> None:
             pass
 
 
-update_registered_conversions()
+_update_registered_conversions()