+++ qbraid/transpiler/converter.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 import warnings
 from copy import deepcopy
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Any, Callable, Optional
 
 from qbraid_core._import import LazyLoader
 
@@ -110,7 +110,17 @@ def transpile(
 
     for path in paths:
         path_details = _get_path_from_bound_methods(path)
-        temp_program = deepcopy(program)
+        try:
+            temp_program = deepcopy(program)
+        except RecursionError as err:
+            logger.warning(
+                "Deepcopy failed due to a %s, likely caused by the internal structure of "
+                "the %s object. Continuing execution, but any subsequent errors during "
+                "transpilation may be unclear or misleading due to potential side effects.",
+                type(err).__name__,
+                type(program),
+            )
+            temp_program = program
         try:
             for convert_func in path:
                 try:
@@ -151,3 +161,36 @@ def transpile(
             else "."
         )
     )
+
+
+def chain_calls(func: Callable[[Any, Any], Any], initial_value, *args, **kwargs) -> Any:
+    """
+    Apply a function iteratively over a sequence of arguments.
+
+    Args:
+        func (Callable[[Any, Any], Any]): The function to apply.
+        initial_value (Any): The initial input to the first call.
+        *args (Any): The sequence of arguments for subsequent calls.
+        **kwargs (Any): Additional keyword arguments for the function.
+
+    Returns:
+        Any: The result of the final call.
+    """
+    result = initial_value
+    for arg in args:
+        result = func(result, arg, **kwargs)
+    return result
+
+
+def translate(program, *targets, **kwargs):
+    """
+    Chain multiple transpile calls on a program.
+
+    Args:
+        program: The quantum program to transpile.
+        *targets: A sequence of target formats.
+
+    Returns:
+        The final transpiled program.
+    """
+    return chain_calls(transpile, program, *targets, **kwargs)