+++ qbraid/transpiler/graph.py
@@ -71,6 +71,15 @@ def _get_path_from_bound_methods(bound_methods: list[Callable[..., Any]]) -> str
     return " -> ".join(path)
 
 
+def parse_conversion_path(conv_repr: str) -> list[tuple[str, str]]:
+    """Parse a conversion path string into a list of conversion tuples."""
+    components = conv_repr.split(" -> ")
+    conversions = []
+    for i in range(len(components) - 1):
+        conversions.append((components[i], components[i + 1]))
+    return conversions
+
+
 # pylint: disable-next=too-many-public-methods
 class ConversionGraph(rx.PyDiGraph):
     """
@@ -86,7 +95,7 @@ class ConversionGraph(rx.PyDiGraph):
         self,
         conversions: Optional[list[Conversion]] = None,
         require_native: bool = False,
-        include_isolated: bool = False,
+        include_isolated: bool = True,
         edge_bias: Optional[float] = None,
         nodes: Optional[Union[list[str], set[str]]] = None,
     ):
@@ -99,7 +108,7 @@ class ConversionGraph(rx.PyDiGraph):
             require_native (bool): If True, only include "native" conversion functions.
                 Defaults to False.
             include_isolated (bool): If True, includes all registered program type aliases, even
-                those that are not connected to any other nodes in the graph. Defaults to False.
+                those that are not connected to any other nodes in the graph. Defaults to True.
             edge_bias (float, optional): Factor used to fine-tune the edge weight calculations
                 and modify the decision thresholds for pathfinding. Defaults to 0.25 to prioritize
                 shorter paths. For example, a bias of 0.25 slightly favors a single conversion at
@@ -303,7 +312,7 @@ class ConversionGraph(rx.PyDiGraph):
             list of list of Callable: The top shortest conversion paths.
 
         Raises:
-            ValueError: If no path is found between source and target.
+            ConversionPathNotFoundError: If no path is found between source and target.
         """
         all_paths = rx.all_simple_paths(
             self, self._node_alias_id_map[source], self._node_alias_id_map[target]